Put the scorefile in ${VARBASE}/games instead of under ${PREFIX}/share. It is still mode 666, which is wrong (XXX), but this way:    - the package will actually install again;    - the score file is not overwritten every time you recompile/reinstall.
There's also an .if 0 block in the makefile that can be turned on to
install the game setgid so the score file can be mode 664. I'm not
comfortable turning this on without doing at least a basic audit, and
I don't have energy for that at the moment; but it's there and it
works if anyone feels inclined to make that tradeoff locally.

An alternative approach would be to write and install a setgid
scorefile helper program, but that's not really any improvement over
making the scores world writeable

PKGREVISION -> 8.