Update to 20111006. Changes: Provide function in libiscsi to return the compile-time limit on number of targets. Use this function in the initiator, to ensure that the same limit is used throughout (i.e. stop you attempting to compile the initiator with a different limit to the library - the initiator uses the library for most of its functionality).
If truncating the number of targets in the initiator, truncate it to the
correct amount, not half of what it should be.

Obey RFC3720 when determining targets. SendTargets=All MUST be supported
on a discovery session, and MUST NOT be supported on an operational session.
Previously, SendTargets=All was used both in the initial discovery session
and the later session. (from Daisuke Aoyama - author of istgt).