Import p5-Switch-2.16 as lang/p5-Switch.
Switch.pm provides the syntax and semantics for an explicit case mechanism for
Perl.  The syntax is minimal, introducing only the keywords C<switch> and
C<case> and conforming to the general pattern of existing Perl control
structures.  The semantics are particularly rich, allowing any one (or more) of
early 30 forms of matching to be used when comparing a switch value with its
various cases.