Added multimedia/adobe-flash-plugin11 version 11.0.1.152
This is based on Ryo ONODERA's wip package with some pkglint and MESSAGE
shuffling from me (any issues seen are likely to be due to that rather
than the original package).

Netscape compatible plugin for Adobe Flash player.  This package
contains a plugin that enables web browsers to render the Flash
format.

The nspluginwrapper option (enabled by default on non Linux platforms)
permits the use of the the Linux x86 flash player in Firefox browsers
on other x86 systems and x86_64 systems.

For NetBSD this package requires emul.linux.kern.osrelease = 2.6.18
or later, which means it will not work on NetBSD 5.x or older
systems.