Update to 0.99.20.
0.99.20:

The primary focus of this release is a fix of SEGV regression in
ospfd, which was introduced in 0.99.19. It also features a series of
minor improvements, including better RFC compliance in bgpd, better
support of FreeBSD and some enhancements to isisd.

0.99.19:

This release provides security fixes, which address assorted
vulnerabilities in bgpd, ospfd and ospf6d (CVE-2011-3323,
CVE-2011-3324, CVE-2011-3325, CVE-2011-3326 and CVE-2011-3327).

0.99.18:

This release fixes 2 denial of services in bgpd, which can be remotely
triggered by malformed AS-Pathlimit or Extended-Community
attributes. These issues have been assigned CVE-2010-1674 and
CVE-2010-1675. Support for AS-Pathlimit has been removed with this
release.

The release includes a number of bug-fixes and enhancements, primarily
for ospfd, ospf6d and bgpd.